[views] renamed to [span] and related changes
Because the `span` type is new in C++20, changing the stable name now does not break previous standards. There is no **[span]** stable name in C++17 nor in the C++20 WD. The same is true for the other stable name changes made here. The section header **Views** was changed to **Class template span**. The remaining changes modify the numbering and move certain text elements to match the style used in the Standard.